+++ index.html
@@ -59,6 +59,6 @@
     </div>
   </div>
   <script src="script.js"></script>
-  <!-- <script src="example.js"></script> -->
+  
 </body>
 </html>
